pyproject.toml
[tool.poetry]
name = "dji3d"
version = "0.1.0"
description = "DJI3D is a tool for graphing 3d positional data extracted from DJI drone telemetry"
authors = ["Evan Pratten <ewpratten@gmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
ffmpeg-python = "^0.2.0"
matplotlib = "^3.3.4"
pysrt = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
